fix: move Interface definition in common module
