+++ packages/core/src/browser/quick-open/quick-open-service.ts
@@ -17,75 +17,20 @@
 import { injectable } from 'inversify';
 import { QuickOpenModel } from './quick-open-model';
 import { MessageType } from '../../common/message-service-protocol';
-import { HideReason } from '../../common/quick-pick-service';
+import * as common from '../../common/quick-open-service';
 
-export type QuickOpenOptions = Partial<QuickOpenOptions.Resolved>;
-export namespace QuickOpenOptions {
-    export interface FuzzyMatchOptions {
-        /**
-         * Default: `false`
-         */
-        enableSeparateSubstringMatching?: boolean
-    }
-    export interface Resolved {
-        readonly prefix: string;
-        readonly placeholder: string;
-        readonly ignoreFocusOut: boolean;
-
-        readonly fuzzyMatchLabel: boolean | FuzzyMatchOptions;
-        readonly fuzzyMatchDetail: boolean | FuzzyMatchOptions;
-        readonly fuzzyMatchDescription: boolean | FuzzyMatchOptions;
-        readonly fuzzySort: boolean;
-
-        /** The amount of first symbols to be ignored by quick open widget (e.g. don't affect matching). */
-        readonly skipPrefix: number;
-
-        /**
-         * Whether to display the items that don't have any highlight.
-         */
-        readonly showItemsWithoutHighlight: boolean;
-
-        /**
-         * `true` if the quick open widget provides a way for the user to securely enter a password.
-         * Otherwise, `false`.
-         */
-        readonly password: boolean;
-
-        selectIndex(lookFor: string): number;
-
-        onClose(canceled: boolean): void;
-    }
-    export const defaultOptions: Resolved = Object.freeze({
-        prefix: '',
-        placeholder: '',
-        ignoreFocusOut: false,
-
-        fuzzyMatchLabel: false,
-        fuzzyMatchDetail: false,
-        fuzzyMatchDescription: false,
-        fuzzySort: false,
-
-        skipPrefix: 0,
-
-        showItemsWithoutHighlight: false,
-        password: false,
-
-        onClose: () => { /* no-op*/ },
-
-        selectIndex: () => -1
-    });
-    export function resolve(options: QuickOpenOptions = {}, source: Resolved = defaultOptions): Resolved {
-        return Object.assign({}, source, options);
-    }
-}
+/**
+ * @deprecated import from `@theia/core/lib/common/quick-open-service` instead
+ */
+export { QuickOpenOptions } from '../../common/quick-open-service';
 
 @injectable()
 export class QuickOpenService {
     /**
      * It should be implemented by an extension, e.g. by the monaco extension.
      */
-    open(model: QuickOpenModel, options?: QuickOpenOptions): void { }
-    hide(reason?: HideReason): void { }
+    open(model: QuickOpenModel, options?: common.QuickOpenOptions): void { }
+    hide(reason?: common.QuickOpenHideReason): void { }
     showDecoration(type: MessageType): void { }
     hideDecoration(): void { }
 }
