+++ packages/core/src/browser/quick-open/quick-pick-service-impl.ts
@@ -17,7 +17,8 @@
 import { injectable, inject } from 'inversify';
 import { QuickOpenItem, QuickOpenMode, QuickOpenGroupItem, QuickOpenItemOptions } from './quick-open-model';
 import { QuickOpenService } from './quick-open-service';
-import { QuickPickService, QuickPickOptions, QuickPickItem, QuickPickSeparator, QuickPickValue, HideReason } from '../../common/quick-pick-service';
+import { QuickPickService, QuickPickOptions, QuickPickItem, QuickPickSeparator, QuickPickValue } from '../../common/quick-pick-service';
+import { QuickOpenHideReason } from '../../common/quick-open-service';
 
 @injectable()
 export class QuickPickServiceImpl implements QuickPickService {
@@ -84,7 +85,7 @@ export class QuickPickServiceImpl implements QuickPickService {
         };
     }
 
-    hide(reason?: HideReason): void {
+    hide(reason?: QuickOpenHideReason): void {
         this.quickOpenService.hide(reason);
     }
