+++ packages/monaco/src/browser/monaco-quick-open-service.ts
@@ -24,7 +24,7 @@ import {
 import { KEY_CODE_MAP } from './monaco-keycode-map';
 import { ContextKey } from '@theia/core/lib/browser/context-key-service';
 import { MonacoContextKeyService } from './monaco-context-key-service';
-import { HideReason } from '@theia/core/lib/common/quick-pick-service';
+import { QuickOpenHideReason } from '@theia/core/lib/common/quick-open-service';
 
 export interface MonacoQuickOpenControllerOpts extends monaco.quickOpen.IQuickOpenControllerOpts {
     readonly prefix?: string;
@@ -72,16 +72,16 @@ export class MonacoQuickOpenService extends QuickOpenService {
         this.internalOpen(new MonacoQuickOpenControllerOptsImpl(model, this.keybindingRegistry, options));
     }
 
-    hide(reason?: HideReason): void {
+    hide(reason?: QuickOpenHideReason): void {
         let hideReason: monaco.quickOpen.HideReason | undefined;
         switch (reason) {
-            case HideReason.ELEMENT_SELECTED:
+            case QuickOpenHideReason.ELEMENT_SELECTED:
                 hideReason = monaco.quickOpen.HideReason.ELEMENT_SELECTED;
                 break;
-            case HideReason.FOCUS_LOST:
+            case QuickOpenHideReason.FOCUS_LOST:
                 hideReason = monaco.quickOpen.HideReason.FOCUS_LOST;
                 break;
-            case HideReason.CANCELED:
+            case QuickOpenHideReason.CANCELED:
                 hideReason = monaco.quickOpen.HideReason.CANCELED;
                 break;
         }
